requirements.txt
# Carbox - JAX-Accelerated Chemical Kinetics Simulation Framework
# Core JAX ecosystem for differentiable computing
jax>=0.4.20
jaxlib>=0.4.20
equinox>=0.11.0
diffrax>=0.4.0

# Scientific computing
numpy>=1.24.0
scipy>=1.10.0
pandas>=2.0.0

# Visualization
matplotlib>=3.7.0

# Utilities
tqdm>=4.65.0
pyyaml>=6.0
